requirements.txt
streamlit
ultralytics
easyocr
numpy
pillow
